Fix mod compilation on non windows platform (!132)
The mod compilation service is assuming the current running platform is windows. It always replaces the import statements with the windows version of the pkg snapshot path.

This breaks loading basically any mods when running on Linux.

I have added the os check and used the correct snapshot path according to the system.

![image](/attachments/4a4dc6fb-8caf-4df8-b6dd-5da40bc10807)

Co-authored-by: Sky Z <qe201020335@sina.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT-AKI/Server/pulls/132
Co-authored-by: qe201020335 <qe201020335@noreply.dev.sp-tarkov.com>
Co-committed-by: qe201020335 <qe201020335@noreply.dev.sp-tarkov.com>