Fix asset hash file generation on non-windows platforms (!137)
The asset hash file path in the gulpfile is hard coded with Windows style path delimiter (`\`). When building on linux, it will generate a hash file with this monstrous file name.
![image](/attachments/2c44f9ce-030b-4671-8bf5-a1bb7af9dac4)

This file name is not only wrong in which the server will not be able to check asset integrity, it will also cause mysterious issues with tools.

For example, pwsh can't handle this at all and commands will fail when touching this file.
![image](/attachments/38a43ee0-deb6-4190-a470-19a7e8beea1d)

A path like `xxx\yyy\zzz/aaa` works on windows, but `xxx/yyy/zzz\aaa` DOES NOT work on linux!

Co-authored-by: Sky Z <qe201020335@sina.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT-AKI/Server/pulls/137
Co-authored-by: qe201020335 <qe201020335@noreply.dev.sp-tarkov.com>
Co-committed-by: qe201020335 <qe201020335@noreply.dev.sp-tarkov.com>