Fix Bonuses in various Services using Find and not filter. (!393)
Fix for SPT/Issues#743
FuelConsumptionBonus only looked for the first value, not all Bonuses a PMC has.
TODO: Find a way for the Broken Wall to not give too much bonus when a player mops the floor in between the stages. Resulting with too much negative bonus once the broken wall is completed.

Co-authored-by: Dev <dev@dev.sp-tarkov.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT/Server/pulls/393
Co-authored-by: TetrisGG <dmiller2606@gmail.com>
Co-committed-by: TetrisGG <dmiller2606@gmail.com>