Database Import Progress Bar (#987)
Waffle wrote a PR (#986) that shows a progress bar when the database is
populating. I pulled it into a local branch and made some adjustments.

I've tested this on my end and it's working okay. In the compiled
project, and directly in terminals (that *I have installed*) it looks
great, but in VSCode's built in terminal you can see escape characters
at the beginning of the line:


![image](https://github.com/user-attachments/assets/c547a675-5331-44f3-a44e-940edaba6fe8)

I can't figure out why, or how to get rid of it, but it doesn't really
bother me much.

---------

Co-authored-by: waffle-lord <76401815+waffle-lord@users.noreply.github.com>