Node Version Validation
This introduces a node version validation check on server start-up. The expected node version is pulled from the `package.json`. Locally, the expected version is pulled directly. The build script writes the expected node version into the `build.json` file, which is used in the built environment. When an invalid match is made the server start-up halts.